test(STRK-246): dedupe slice helpers into test-helpers.js + reuse bulk-edit setup (PR #1337 Codacy)
